+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/kotlin/_internal/createAndConfigureHubdleKotlinExtensions.kt
@@ -12,6 +12,7 @@ import com.javiersc.hubdle.project.extensions.kotlin.android.library.features.Hu
 import com.javiersc.hubdle.project.extensions.kotlin.compiler.options.HubdleKotlinCompilerOptionsExtension
 import com.javiersc.hubdle.project.extensions.kotlin.features.HubdleKotlinFeaturesExtension
 import com.javiersc.hubdle.project.extensions.kotlin.features.shared.HubdleKotlinAtomicfuFeatureExtension
+import com.javiersc.hubdle.project.extensions.kotlin.features.shared.HubdleKotlinBuildKonfigFeatureExtension
 import com.javiersc.hubdle.project.extensions.kotlin.features.shared.HubdleKotlinComposeFeatureExtension
 import com.javiersc.hubdle.project.extensions.kotlin.features.shared.HubdleKotlinContextReceiversFeatureExtension
 import com.javiersc.hubdle.project.extensions.kotlin.features.shared.HubdleKotlinCoroutinesFeatureExtension
@@ -76,6 +77,7 @@ private fun HubdleState.createKotlinFeatureExtensions() {
     createExtension<HubdleKotlinFeaturesExtension> {
         createExtension<HubdleKotlinCompilerPluginFeatureExtension>()
         createExtension<HubdleKotlinAtomicfuFeatureExtension>()
+        createExtension<HubdleKotlinBuildKonfigFeatureExtension>()
         createExtension<HubdleKotlinComposeFeatureExtension>()
         createExtension<HubdleKotlinContextReceiversFeatureExtension>()
         createExtension<HubdleKotlinCoroutinesFeatureExtension>()